pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "intent-compiler"
version = "0.1.0"
description = "Deterministic Markdown → structured artifacts. Intent compiler for structured protocols."
authors = [{ name = "Protocol Team" }]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "click>=8.0",
    "markdown-it-py>=3.0",
    "jinja2>=3.0"
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0"
]

[project.scripts]
intent = "intent_compiler.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
